src/api/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
fastapi==0.109.0
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
joblib==1.3.2
Levenshtein==0.23.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
passlib==1.7.4
pluggy==1.3.0
psycopg2-binary==2.9.9
pyarrow==15.0.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pytest==7.4.4
python-dateutil==2.8.2
python-Levenshtein==0.23.0
python-multipart==0.0.6
pytz==2023.3.post1
rapidfuzz==3.6.1
requests==2.31.0
scikit-learn==1.4.0
scikit-surprise==1.1.3
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.27.0
mlflow
prometheus-fastapi-instrumentator
prometheus-client
flake8